more code cleanup
